restructure Windows filesystem inspection logic
